feat: worktree cleanup wip
Sometimes branches and worktrees don't get cleaned up, and that's ok since a user might want to continue some work where it was left off.
What I propose instead is adding a dedicated modal in the frontend to manage worktrees and branches.
The user should be able to see the task - attemps => worktrees relationship.
The task should indicate the task status, like if it was merged and which commit (clickable to open commit details).
When clicking on an attempt, the task attempt should be displayed, like clicking on a recent task loads the last available attempt.
Then, the user should be able to delete an attempt, meaning deleting that worktree.
This action should delete the worktree but keep the branch. The worktree card in the UI should not become "ghostly" indicating that the worktree was deleted, but the branch remains.
Then, the user can press the same button to delete the branch as well.
If the user shift+clicked the button (with icon changed to danger triangle) then the worktree and branch deletion will happen in one click instead of 2 separate clicks.

In any case, for all deletion operations, the user should be asked for confirmation.