feat: restart session
We change the idea of compacting from something the backend handles transparently to instead be supported by the ability of asking the backend to create a new session with the summary from the previous process.
This allows the frontend to ask the agent to compact its context, and then signal the backend to start a new session with the returned summary and the followup request.